requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11
django-bootstrap3==8.2.2
django-crispy-forms==1.6.1
django-filter==1.0.4
django-graphos==0.3.41
django-tables2==1.6.1
gunicorn==19.7.1
pytz==2017.2
static3==0.7.0
psycopg2>=2.6.1
